Improve wording.
It's slightly strange to say that I don't want to hide behind the phrase 'on request' at the top of the page, and then I use it later in the page.  In reality, the only reason I say that it 'can' be done is that I need to generate it specifically for that person and have it emailed to them by the University's system.